requirements.txt
# Core Libraries
python-dotenv

# File Parsing
PyMuPDF
python-docx

# NLP & AI
spacy
langchain
langchain-community 
langchain-google-genai # Or langchain-openai, etc.
sentence-transformers # For embeddings

# Vector Store
chromadb

# Web Application
streamlit
pandas 